Fix serious bug in send/receive loop, better ACK
Discovered that the "send all bytes just read" loop in sendFile/getFile was
seriously broken. Seemed to work on tcp but not UDT - very weird because it
was really broken:

    // read a new block of data
    nWritten = 0, nRead = read(..., n);

    // the idea: keep on looping until all bytes just read are sent
    while( nWritten<nRead ) {
        // attempt a single writ
        thiswrite = write(..., nRead - nWritten);

        // check if anything written, break loop if not (snipped)

        // update loop counter(s) <===== BUT VERY WRONG!!!!!!
        nWritten += thiswrite;  // this is OK
        nRead    -= thiswrite;  // THIS IS ABSOLUTELY NOT OK!!!
    }

With TCP sockets this worked Just Fine, with UDT very much not (both in
blocking mode, by the way)

What happens is that a blocking TCP write only returns if all the bytes are
sent so the loop is only executed once and all bytes are transferred.

With UDT, apparently, it doesn't write all bytes at once, meaning that
"thiswrite" < nRead, meaning that both nWritten and nRead got updated,
breaking the loop termination condition :-(  and not sending all bytes
across that were physically read.

With this commit that is now fixed and the code behaves correctly for both
TCP, UDT and blocking as well as non-blocking sockets.

Secondly, the ACK is unconditionally sent and only waited for if the
destination connection still seems to be alive. If the code exited the
previous loop because of an error on the destination, there is no point in
(trying) to send the ACK.