Fix SIGSEGV on print, compile errors and ^C handling
* Newer compilerts = new compiler ishoos

* Avoid sigsegv/exception on fmtTime(...), fmt1000() and friends:

The mk_format() function returned a lambda that used perfect
forwarding of mk_format() arguments - references to temporaries in this
case. Could not capture by copying since some of the args could be
std::facet / std::numpunct which are non-copyable by design.

Fixed by processing the arguments to mk_format() into a ref counted
pointer to local variable in the closure which can be captured by value making the variables reachable from inside the lambda w/o SIGSEGV on access.

Took a while but now the printin' of values is both prettyfied and it don't SIGSEGV no more.

* Gracefully handle ^C

The etc client should now gracefully handle ^C; the output now shows if the transfer finished as well as how many bytes were transferred in how much time.

* make -v work intuitively

Also the '-v' command line option was counterintuitive; passing '-v' silenced the output.
So now it defaults to quiet and '-v' makes the output more verbose.