Avoid sigsegv/exception on fmtTime(...)
The SIGSEGV in fmt1000() was fixed, what remains is an exception thrown by fmtTime(...).
The thousands' value for fmtTime() seems to be either uninitialized or erroneously initialized - triggering runtime_error: thousands's value should be > 1

Pending proper fix avoid pretty printing the time for now so transfer of multple files will continue to run.