Gracefully handle ^C, make -v work intuitively
The etc client should now gracefully handle ^C; the output now shows if the transfer finished as well as how many bytes were transferred in how much time.
The beautification of the output is still to do.

Also the '-v' command line option was counterintuitive; passing '-v' silenced the output.
So now it defaults to quiet and '-v' makes the output more verbose.