this should close issue #7
File existing on the remote end is not translated into a magic error
code that makes the client not attempt N times but give up immediately
as this is a user error.

It might even work with protocol version 0 daemons - if the server admin
has not switched their system language to one where the error message
for EEXIST does not contain the literal string "File exists" (without
the quotes).

Protocol version 1+ daemons do not leave it up to strerror(3) to come up
with the error message but literally return "File exists" irrespective
of locale.