Bugfix for 1.0
During local testing with multiple data channels found that the code
parsing the received data channels from the server was off by a comma.

Error message when trying to transfer to a server with > 1 data
channels:

2021-06-02 14:03:53.87: ETDServerWrapper: sending reply 'ERR assertion error: src/etdc_etdserver.cc:72 [std::regex_match(s, fields, rxSockName)] The string '<udt6/localhost:8008/mss=1500,max-bw=-1>,' is not a valid data address designator'

Fixed, tested to work with one as well as more than one data channels
and bumped version to 1.0.1 in Makefile.