Use unused return value from write(2)
Newer compiler than I have access to is more strict and complains that
return value from ::write(2) wasn't used. Now it is, although probably,
if the error is triggered, nothing useful can be done since the process
is daemonized, stdin/stdout have been redirected to /dev/null, and the
error involves a failure to write to stderr ... so any error messages
will likely also not be written to stderr on account an error happened
whilst attempting that with a previous message ...

But hey, the compiler's happy now!