Provide possible fix + more error details
It could be that ::mkdir() doesn't set errno and thus the assertion
could be done on a value that ::mkdir() didn't actually set. Fixed by
always (re)initializing errno before executing and asserting.

To aid debugging the errno and error text have been added to the output;
in case the assertion fails the message didn't mention anything useful
as to why it might have failed.